Lots of housekeeping
finally got the inner_nav to work properly online. for some reason the $layout_context would not set when it was defined at the top of the page. I moved it down to where the inner_nav was "required" and it works now. also did some cleaning up with the share_project.php page.